kernel_10_26_1_18/jb-mr2-t40-r1-10.26.1.18-20131217-1558-foss/mydroid/external/webkit/LayoutTests/platform/android/fast/workers/shared-worker-script-error-expected.txt

CONSOLE MESSAGE: line 12: Uncaught ReferenceError: generateError is not defined
Test SharedWorker script error handling functionality. Should print a series of PASS messages, followed with DONE.

SUCCESS: unhandled error generated
SUCCESS: error handled via onerror: Uncaught ReferenceError: generateError is not defined
DONE
